ta quase

+++ src/app/account/logout/logout.component.ts
@@ -1,18 +1,26 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { AlertService } from 'src/app/services/alert.service';
 
 @Component({
   selector: 'app-logout',
   templateUrl: './logout.component.html',
-  styleUrls: ['./logout.component.css']
+  styleUrls: ['./logout.component.css'],
 })
 export class LogoutComponent implements OnInit {
+  constructor(private router: Router, private alert: AlertService) {}
 
-  constructor(private router: Router) { }
-
-  ngOnInit(): void {
-    window.localStorage.clear();
-    this.router.navigate(['']);
+  sair: any;
+  async logout() {
+    this.sair = await this.alert.interrogation('Deseja realmente sair?');
+    if (this.sair.status) {
+      window.localStorage.clear();
+      this.router.navigate(['']);
+    } else {
+      this.router.navigate(['']);
+    }
   }
-
+  ngOnInit(): void {
+    this.logout();
+  }
 }
